+++ benchmark/corpus/cases/go-md5-real/main.go
@@ -1,0 +1,8 @@
+package main
+
+import "crypto/md5"
+
+func f() []byte {
+	h := md5.New()
+	return h.Sum(nil)
+}