+++ benchmark/corpus/cases/go-string-md5/main.go
@@ -1,0 +1,5 @@
+package main
+
+func f() string {
+	return "md5.New() disabled"
+}