+++ benchmark/corpus/cases/java-des-real/B.java
@@ -1,0 +1,7 @@
+import javax.crypto.Cipher;
+
+class B {
+  void f() throws Exception {
+    Cipher c = Cipher.getInstance("DES/ECB/PKCS5Padding");
+  }
+}