+++ benchmark/corpus/cases/java-string-md5/D.java
@@ -1,0 +1,5 @@
+class D {
+  void f() {
+    String m = "never call MessageDigest.getInstance('MD5')";
+  }
+}