+++ benchmark/corpus/cases/js-comment-md5/app.js
@@ -1,0 +1,3 @@
+function f() {
+  // crypto.createHash('md5') is forbidden
+}