+++ benchmark/corpus/cases/js-md5-real/app.js
@@ -1,0 +1,3 @@
+const crypto = require('crypto');
+
+const h = crypto.createHash('md5');