+++ benchmark/corpus/cases/php-md5-real/index.php
@@ -1,0 +1,2 @@
+<?php
+$hash = md5($password);